Implemented an EnhancedNumber class
the EnhancedNumber class leverages any function calls against a number to cache meta data about the number that is useful in subsequent function calls.

For example, when calculating the Divisors of a number, the factorsOf, getCountDivisors and the Sieve for the number are all calculated.

With those results cached, the subsequent calls to other or the same function will be much faster, since there is no calculation to be done, simply cache retrieval.

For best results, the user should use an EnhancedNumber when calling functions, otherwise the meta data is lost at the point of returning the value of the function called, but if you call a function using an EnhancedNumber, that meta data is preserved.